@@ -1,5 +1,5 @@
 class TagsController < ResourceController
-  before_filter :authenticate_user!
+  before_action :authenticate_user!
 
   def index
     @tomatoes_by_tag = Rails.cache.fetch("tomatoes_by_tag_#{current_user.id}", expires_in: 1.day) do
