@@ -7,5 +7,5 @@
 //= require jquery
 //= require jquery_ujs
 //= require jquery.cookie
-//= require jquery.color
+//= require jquery.color-2.1.2
 //= require time_zone
